move shareable code into own file
