+++ src/utils/email.py
@@ -48,3 +48,16 @@ class Emailer:
     @staticmethod
     def apply_template_values(template, values):
         return template.format(**values)
+
+
+def send_email(emailer,
+               report,
+               template_name,
+               template_values,
+               config,
+               additional_recipients):
+
+    recipients = report.success_email_recipients + additional_recipients
+    emailer.send(config.smtp_from, recipients,
+                 report.success_email_subject, template_name,
+                 template_values)
